pyproject.toml
[project]
name = "git-diff-critique"
version = "0.1.0"
description = "An agent that reviews git diffs and provides actionable, senior-level code review recommendations using Google ADK."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.3.0",
    "pytest>=8.4.0",
    "ruff>=0.11.13",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
